Reproduce failure
It appears that when we have a file in `extra_srcs` that is generated by
a rule like `write_file`, the file is not correctly tracked as a
dependency. This doesn't seem to fail if either the file is a source
file (not generated) or if we have a generated file in `srcs`. That
seems to indicate that the issue is with `extra_srcs`.